+++ BurningKnight/changelog.md
@@ -2,3 +2,5 @@
 
 * Changelog started
 * Version now supports string name
+* Don't let to swap weapons, if you don't have another one
+* Interact on gamepad should be X